cmd/wasm-client: move shared code to core package
